Refactor processed-payments tests: remove redundant/impossible tests, add stale recovery coverage, split by concern
The original test file had several quality issues:

1. `beforeEach` created 3 attendees for every test even though most
   tests only need 1, adding unnecessary latency to the entire suite.

2. The `two-phase session processing` describe block retested the same
   behaviors already covered by `reserveSession` and `finalizeSession`
   individually—pure duplication with no added signal.

3. The `stale reservation recovery` block duplicated reserveSession
   tests (fresh reservation stays blocked; finalized reservation stays
   blocked) and included a constant-value assertion
   (`STALE_RESERVATION_MS is 5 minutes`) that tests the import, not
   behavior.

4. The `reserveSession race condition recovery` test simulated a record
   vanishing between an INSERT and a SELECT by stubbing `getDb().execute`.
   This is exactly the "system randomly crashes between two lines"
   scenario our test quality standards prohibit, and it exercised a
   recursive code path explicitly labelled "Shouldn't happen in practice".

5. `getProcessedAttendeeId` had two tests asserting the same thing—
   one via the `processSession` helper, one via direct reserve+finalize—
   with no difference in what was being verified.

6. `isReservationStale` had a "very old timestamp" test redundant with
   the "over threshold" boundary test.

7. `env.test.ts` patched `globalThis.Deno` to `undefined` to test a
   defensive throw that can never fire in either of our runtime
   environments (Deno local dev, Bunny Edge prod).

Changes made:
- Split the 555-line file into `processed-payments/locking.test.ts`
  (session reserve/finalize lifecycle) and `processed-payments/staleness.test.ts`
  (isReservationStale, deleteStaleReservation, deleteAllStaleReservations).
- Reduced `beforeEach` to create a single attendee; the one test that
  needs multiple creates them inline.
- Removed all the redundant/impossible tests listed above.
- Added a genuine missing test: `reserveSession` recovering a stale
  unfinalized reservation—a key production behaviour (crashed checkout
  process, retry after 5 min) that had zero coverage.
- Removed the `globalThis.Deno = undefined` test from `env.test.ts`.
